+++ apps/block_scout_web/lib/block_scout_web/controllers/account/api/v2/fallback_controller.ex
@@ -146,6 +146,13 @@ defmodule BlockScoutWeb.Account.API.V2.FallbackController do
     |> render(:message, %{message: "No Bearer token"})
   end
 
+  def call(conn, {:enabled, false}) do
+    conn
+    |> put_status(:not_found)
+    |> put_view(UserView)
+    |> render(:message, %{message: "This endpoint is not configured"})
+  end
+
   defp unauthorized_error(%{email_verified: false, email: email}) do
     %{message: "Unverified email", email: email}
   end